fixed noop constraints
